Updated cli, backup, and rm_host

+++ src/rm_known_host/rm_host.py
@@ -1,5 +1,5 @@
 
-def delete_line(stringToMatch, file_in):
+def delete_line(stringToMatch, file_in, interactive=False):
     print(f"Remove {stringToMatch} from {file_in}")
     output = []
     counter = 0
@@ -9,16 +9,19 @@ def delete_line(stringToMatch, file_in):
                 output.append(line)
             else:
                 counter += 1
+    
+    confirm = 'Y'
+    if counter == 0:
+        print(f"String {stringToMatch} was not found!")
 
-
-    if counter > 0:
+    if interactive and counter > 0:
+        confirm = input(f"The known_host file has {counter} matches to {stringToMatch} Continue(Y/N):")
+    
+    if confirm.upper() == 'Y' and counter > 0:
         file_out = open(file_in, 'w')
         file_out.writelines(output)
         file_out.close()
-    else:
-        print(f"String {stringToMatch} was not found!")
+        print(f"Removed lines containing {stringToMatch} from {file_in}.")
+        
 
-
-delete_line('3.9', '/home/cloud_user/.ssh/known_hosts')
-
-#new_output = open()
+# delete_line('18.2', '/home/cloud_user/.ssh/known_hosts', True)
